+++ hooks/postdown.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# check to see if this is running as github action
+if [[ ! -z "$CI" ]]; then
+   echo "Running on github, skipping postdown hook. Exiting."
+   exit
+fi
+
 if [ -f "${HOME}/.kube/${AZURE_AKS_CLUSTER_NAME}" ]; then
    rm ${HOME}/.kube/${AZURE_AKS_CLUSTER_NAME}
 fi
