+++ hooks/predown.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# check to see if this is running as github action
+if [[ ! -z "$CI" ]]; then
+   echo "Running on github, skipping predown hook. Exiting."
+   exit
+fi
+
 if kubectl config get-contexts ${AZURE_AKS_CLUSTER_NAME} -o name 2>&1| grep ${AZURE_AKS_CLUSTER_NAME}; then
    echo "Remove previous kubeconfig context ${AZURE_AKS_CLUSTER_NAME}"
    kubectl config delete-context ${AZURE_AKS_CLUSTER_NAME} --kubeconfig=${HOME}/.kube/config
